Add vector store apis for assistants
Add functions to create and read vector stores and vector store file batches. Add a separate helper function to handle uploading files, then associating them with the vector store.

The helper function to upload and poll was more for my personal needs, so can remove it for this PR if it is not something you want to include in the lib.

I have also updated the assistants beta version to v2 by default - can also remove this if needed.